feat(desktop): add generation-host selection to the Sequence header
The chain composer routed purely by model (Auto/capable) with no way to
pin a host, unlike Create's header chip. Extract that chip into a shared
HostChip component (same generateTargetHost contract, DOM, and menu) and
mount it in the Sequence header; chain-limits fetches, durable-job
creation, and job actions now follow a sticky pick, while Auto keeps the
model-aware default that escalates CUDA-only families to Most capable.
