+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **The desktop chain composer is now discoverable from Create.** Sequence authoring existed only behind File → New Chain and the ⌘K palette, so the Create workspace showed no path to it at all. The Create header now carries the same **Single | Sequence** segmented switch as web and iPhone, and the chain composer's hand-rolled toggle was replaced with the shared Mold Studio segmented control so the pair round-trips consistently.
 - **Desktop Sequence now shares Create's chrome and model picker.** Switching Single ↔ Sequence no longer swaps to a differently shaped page: the chain composer uses the same 52px header anatomy (title, summary pill, mode switch), and its plain dropdown was replaced with the Create inspector's model picker — extracted into a shared component — filtered to installed video models. Chain-incapable checkpoints such as two-stage LTX-2 `-dev` exports now appear disabled with the reason instead of silently vanishing from the list.
+- **Desktop Sequence now offers the same host selection as Single.** The chain composer header carries the shared generation-host chip — Auto (least busy), Most capable, or a sticky host, writing the same persisted `generateTargetHost` contract — and chain limits, creation, and job actions follow a sticky pick, while Auto keeps the model-aware default that escalates CUDA-only families to the most capable CUDA machine.
 - **A long chain jobs list no longer crushes the sequence filmstrip.** The chain composer's fixed rows (header, filmstrip, footer, TOML panel) were flex children of the scrolling column and compressed into overlapping slivers once enough jobs accumulated; they now hold their height and the jobs list scrolls instead. The jobs header also replaces the bare `GC` glyph with **Clear inactive** — deleting every job that isn't running or queued, including resumable ones, in one action — and a labeled **Clean up disk** sweep.
 - **The chain TOML panel is now a syntax-highlighted editor.** View as TOML shows the script with token-colored tables, keys, strings, numbers, and comments, and edits can be applied back to the filmstrip through the same parser as Open .toml — a parse error is shown inline without touching the sequence. The File tools menu also dismisses on outside click, Escape, and after picking an action, and the chain start-image picker closes on Escape even when focus sits outside the dialog (WKWebView never focuses clicked buttons).
